% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np_prices.R
\name{np_prices}
\alias{np_prices}
\title{Get the spot prices for a specified price_date.}
\usage{
np_prices(time_unit = c("hourly", "daily", "weekly", "monthly", "yearly"),
  currency = c("EUR", "NOK", "DKK", "SEK"), price_date = Sys.Date() +
  lubridate::days(1), areas = NULL, long_data = FALSE)
}
\arguments{
\item{time_unit}{The time unit to return the data in.}

\item{currency}{Price currency. EUR, NOK, DKK or SEK.}

\item{price_date}{Date class object.}

\item{areas}{Vector with listed areas.}

\item{long_data}{Boolean. Do you want the data returned as a long or wide data.frame?}
}
\description{
Get the spot prices for a specified price_date.
}
\examples{
  # Get the prices for tommorow.
  np_prices("hourly", "EUR")

  # Get the prices for today
  np_prices("hourly", "EUR", Sys.Date())

  # Get a long data frame
  np_prices("hourly", "EUR", long_data = TRUE)

  # Get only the Danish prices in DKK
  np_prices(time_unit = "hourly", currency = "DKK", areas = c("DK1", "DK2"))
}

